+++ src/Types/index.ts
@@ -1,0 +1,7 @@
+export type Categories =
+  | "about"
+  | "education"
+  | "experiences"
+  | "projects"
+  | "skills"
+  | "social"
