+++ src/components/Template/_components/ProfesionalPerfil/_components/Education/_components/Item/index.tsx
@@ -1,21 +1,18 @@
 import { Editable } from "@components/common"
+import { Categories } from "@Types"
 import { Experience_item } from "tree"
 import { Tecnologies } from "../../../common"
-
-type Categories =
-  | "about"
-  | "education"
-  | "experiences"
-  | "projects"
-  | "skills"
-  | "social"
+import { Capa_item } from "../../../_common/Capa_item"
 
 type ItemProps = Experience_item & {
+  id: number
   index: number
   category: Categories
+  remove_item: any // method with 1 parameter number
 }
 
 export const Item = ({
+  id,
   index,
   company,
   profesion,
@@ -23,30 +20,36 @@ export const Item = ({
   summary,
   tecnologies,
   category,
+  remove_item,
 }: ItemProps) => (
-  <div className="flex flex-col gap-[15px]">
-    <div className="flex justify-between font-bold">
-      <div className="flex flex-col gap-1">
-        {/*  */}
-        <div className="max-w-[220px]">
-          {/* company */}
-          <Editable name={`${category}.${index}.company`} value={company} />
-          {/* profesion */}
-          <Editable name={`${category}.${index}.profesion`} value={profesion} />
+  <Capa_item {...{ id, remove_item }}>
+    <>
+      <div className="flex justify-between font-bold">
+        <div className="flex flex-col gap-1">
+          {/*  */}
+          <div className="max-w-[220px]">
+            {/* company */}
+            <Editable name={`${category}.${index}.company`} value={company} />
+            {/* profesion */}
+            <Editable
+              name={`${category}.${index}.profesion`}
+              value={profesion}
+            />
+          </div>
+        </div>
+        {/* date */}
+        <div className="flex gap-1">
+          <Editable name={`${category}.${index}.date.start`} value={start} />
+          <div>-</div>
+          <Editable name={`${category}.${index}.date.end`} value={end} />
         </div>
       </div>
-      {/* date */}
-      <div className="flex gap-1">
-        <Editable name={`${category}.${index}.date.start`} value={start} />
-        <div>-</div>
-        <Editable name={`${category}.${index}.date.end`} value={end} />
-      </div>
-    </div>
 
-    {/* summary */}
-    <Editable name={`${category}.${index}.summary`} value={summary} />
+      {/* summary */}
+      <Editable name={`${category}.${index}.summary`} value={summary} />
 
-    {/* tecnologies */}
-    <Tecnologies {...{ tecnologies, last_index: index }} />
-  </div>
+      {/* tecnologies */}
+      <Tecnologies {...{ tecnologies, last_index: index }} />
+    </>
+  </Capa_item>
 )
