--- src/components/Template/_components/ProfesionalPerfil/_components/Education/_components/items/index.ts
@@ -1,26 +1,0 @@
-export const items = [
-  {
-    id: 1,
-    institute: "ESCUELA COLOMBIANA DE CARRERAS INDUSTRIALES",
-    title: "System Engineering",
-    date: {
-      start: "2012",
-      end: "2014",
-    },
-    summary:
-      "3 semesters of studies to reach the engineering degree. The content of the studies were programming algorithms, functions, algebra, physics, fundamentals of engineering.",
-    tecnologies: ["java"],
-  },
-  {
-    id: 2,
-    institute: "CODER HOUSE",
-    title: "Front End Web Developer",
-    date: {
-      start: "2015",
-      end: "2016",
-    },
-    summary:
-      "Creation and design of web pages with HTML5, JS, CSS, bootstrap, jquery, APIs, animations. As self-educated person I improved my professional skills.",
-    tecnologies: ["HTML", "CSS", "JavaScript"],
-  },
-]
