+++ src/components/Template/_components/ProfesionalPerfil/_components/Experiences/_components/Item/index.tsx
@@ -1,7 +1,7 @@
 import { Editable } from "@components/common"
-import { useState } from "react"
 import { Experience_item } from "tree"
 import { Tecnologies } from "../../../common"
+import { Capa_item } from "../../../_common/Capa_item"
 
 type Categories =
   | "about"
@@ -17,33 +17,6 @@ type ItemProps = Experience_item & {
   remove_item: any // method with 1 parameter number
 }
 
-const useEtwas = (id, remove_item) => {
-  const [show_delete_button, set_show_delete_button] = useState(false)
-
-  const handle_mose_over_item = () => {
-    set_show_delete_button(true)
-  }
-
-  const handle_mose_leave_item = () => {
-    set_show_delete_button(false)
-  }
-
-  const handle_delete_item = () => {
-    remove_item(id)
-  }
-
-  const handlers = {
-    onMouseOver: handle_mose_over_item,
-    onMouseLeave: handle_mose_leave_item,
-  }
-
-  return {
-    handlers,
-    handle_delete_item,
-    show_delete_button,
-  }
-}
-
 export const Item = ({
   id,
   index,
@@ -55,47 +28,36 @@ export const Item = ({
   category = "experiences",
   remove_item,
 }: ItemProps) => {
-  const { handlers, handle_delete_item, show_delete_button } = useEtwas(
-    id,
-    remove_item
-  )
-
   return (
-    <div className="flex flex-col gap-[15px] relative" {...handlers}>
-      {show_delete_button && (
-        <button
-          className="absolute left-1/2 animate-fade-in"
-          onClick={handle_delete_item}
-        >
-          delete
-        </button>
-      )}
-      <div className="flex justify-between font-bold">
-        <div className="flex flex-col gap-1">
-          {/*  */}
-          <div className="max-w-[220px]">
-            {/* company */}
-            <Editable name={`${category}.${index}.company`} value={company} />
-            {/* profesion */}
-            <Editable
-              name={`${category}.${index}.profesion`}
-              value={profesion}
-            />
+    <Capa_item {...{ id, remove_item }}>
+      <>
+        <div className="flex justify-between font-bold">
+          <div className="flex flex-col gap-1">
+            {/*  */}
+            <div className="max-w-[220px]">
+              {/* company */}
+              <Editable name={`${category}.${index}.company`} value={company} />
+              {/* profesion */}
+              <Editable
+                name={`${category}.${index}.profesion`}
+                value={profesion}
+              />
+            </div>
+          </div>
+          {/* date */}
+          <div className="flex gap-1">
+            <Editable name={`${category}.${index}.date.start`} value={start} />
+            -
+            <Editable name={`${category}.${index}.date.end`} value={end} />
           </div>
         </div>
-        {/* date */}
-        <div className="flex gap-1">
-          <Editable name={`${category}.${index}.date.start`} value={start} />
-          -
-          <Editable name={`${category}.${index}.date.end`} value={end} />
-        </div>
-      </div>
 
-      {/* summary */}
-      <Editable name={`${category}.${index}.summary`} value={summary} />
+        {/* summary */}
+        <Editable name={`${category}.${index}.summary`} value={summary} />
 
-      {/* tecnologies */}
-      <Tecnologies {...{ tecnologies, last_index: index }} />
-    </div>
+        {/* tecnologies */}
+        <Tecnologies {...{ tecnologies, last_index: index }} />
+      </>
+    </Capa_item>
   )
 }
