+++ src/components/Template/_components/ProfesionalPerfil/_components/Experiences/_components/index.ts
@@ -1,3 +1,1 @@
 export * from "./Item"
-
-export * from "./items"
