--- src/components/Template/_components/ProfesionalPerfil/_components/Experiences/_components/items/index.ts
@@ -1,95 +1,0 @@
-export const items = [
-  {
-    id: 1,
-    company: "DEV BASE",
-    title: "Leader",
-    date: {
-      start: "2020",
-      end: "2022",
-    },
-    summary:
-      "My responsibilities include leading the Zuro team, which consists of 8 developers, including those responsible for backend, front end, and DevOps.",
-    tecnologies: [
-      "React",
-      "Redux",
-      "TypeScript",
-      "StyleComponents",
-      "Figma",
-      "Jira",
-      "SCRAM",
-    ],
-  },
-  {
-    id: 2,
-    company: "MAKING SENSE",
-    title: "UX Senior Developer",
-    date: {
-      start: "2020",
-      end: "2020",
-    },
-    summary:
-      "My responsibility was to be a connection between Designers and Developers, One of my Skills was to make my opinions known about user experience, design and software. Research in severals ways of Design, develop front end behavior and interaction with back end, decide appropriate approach code and agree for easy and excellent interface with my teams.",
-    tecnologies: [
-      "React",
-      "Redux",
-      "TypeScript",
-      "StyleComponents",
-      "Figma",
-      "Sketch",
-      "Jira",
-      "SCRAM",
-    ],
-  },
-  {
-    id: 3,
-    company: "WENANCE",
-    title: "Front End Engineer",
-    date: {
-      start: "2019",
-      end: "2020",
-    },
-    summary:
-      "Solution thinker and provider. With my engineering capabilities complex projects in the Financial industry will be brought to life. In internal projects client identifying data will be processed. To fulfill the requirents of these projects, applications and multiple languages are used like:",
-    tecnologies: [
-      "React",
-      "Redux",
-      "Node.js",
-      "TypeScript",
-      "CSS",
-      "StyledComponents",
-    ],
-  },
-  {
-    id: 4,
-    company: "EULAS S.A.",
-    title: "UI/UX Developer",
-    date: {
-      start: "2019",
-      end: "2016",
-    },
-    summary:
-      "Maintaining the design and functionality of e-commerce. The business was located in United States of America. Projects were realized with the following companies: Ascena Group, Stadium Goods, Children Place. The following technologies were used for the projects:",
-    tecnologies: [
-      "HTML",
-      "CSS",
-      "JS",
-      "JQuery",
-      "Sass",
-      "Node.js",
-      "Express",
-      "React",
-      "Redux",
-      "Jira",
-      "CMS",
-      "Webpack",
-      "Grunt",
-      "SCRAM",
-    ],
-    examples: [
-      {
-        name: "https://www.louandgrey.com/",
-        path: "https://www.anntaylor.com/",
-      },
-    ],
-  },
-]
